Bug 911987 part 4 - Add onwebkitanimation** and onwebkittransitionend event handlers. r=smaug
Add the onwebkit prefixed event handler attributes of CSS-Animation
and CSS-Transition.

We need to support both content attributes and IDL attributes like
onanimation** attributes.

We should support attribute of lower-case and mixed-case. But currently
Gecko defined as mixed-case only. [1][2]

[1] https://dxr.mozilla.org/mozilla-central/rev/6608e5864780589b25d5421c3d3673ab30c4c318/dom/base/nsContentUtils.cpp#747
[2] https://dxr.mozilla.org/mozilla-central/rev/6608e5864780589b25d5421c3d3673ab30c4c318/dom/events/EventNameList.h#977

So we can't find the related Atom when we set content attributes as lower-case. [3]

[3] https://dxr.mozilla.org/mozilla-central/rev/6608e5864780589b25d5421c3d3673ab30c4c318/dom/base/nsGlobalWindow.h#818

In this patch, I add two Atoms in order to support mixed and lower case.