Bug 1295008 - Hide the previous eyedropper when we request to show a new one; r=miker
The eyedropper can be shown in 2 distinct ways:
- the 'eyedropper' gcli command can be called (from the dev toolboar
  or from the browser devtools menu),
- or the inspector's 'pickColorFromPage' method can be called (from
  the inspector toolbar or from the color-picker in the ruleview).

Before this change, it was possible to show several eyedropper because
these 2 codepaths didn't know about each other.
Now, when executing the gcli command, the inspector's
'cancelPickColorFromPage' method is called to hide it first.
And, when the 'pickColorFromPage' method is called, the 'eyedropper --hide'
gcli command is called too.

This way, there's only one eyedropper shown.

There was also a problem where the gcli command would create a new
eyedropper everytime it was called. This was fixed too, by maintaining a
WeakMap of all eyedroppers opened so far.