Bug 1298472 - part 2 - use #include_next in STL header wrappers for clang-cl; r=glandium
clang-cl's system header detection doesn't kick in if we're using full
paths to the STL headers we need to include.  And if certain headers
aren't recognized as system headers, we get a pile of warnings from
said headers.  Use #include_next to ensure that system headers are
recognized as such, and suppress the warnings.