Bug 1298438 - add modal highlighting stylesheet to each document, not only for each window. r=jaws
The reason this works is because the window object stays the same between location
changes, but the document object changes.
Since stylesheets are applied to documents, not windows, we need to make sure to
inject a stylesheet for each document, otherwise the find toolbar highlights will
look unstyled and wonky.
In bug 1279695 I made a change - to make tests pass and patch landable - to keep
track of all the things (state, mainly range sets) mapped to window objects, because
I hadn't thought of the fact that the content scripts and JSMs that are loaded in
its scope will see many different window objects. As many as there are tabs open,
actually (not counting frames).
The mistake I made there was to keep track of the state flag 'did I already install
the stylesheet?' on a per-window basis, unlike I did before. This patch corrects
this back to the way it was and should be, really.