Bug 1276132 - Rename media.gmp-*.forcevisible to media.gmp-*.visible, and set it when keysystems are enabled. r=spohl
Repurpose the media.gmp-*.forcevisible pref to control whether the
corresponding GMP is visible in the addons manager UI. The pref has to be true
for the GMP to be usable.

The pref is enabled and not hidden when the corresponding EME keysystem is
enabled in the mozconfig.

This means users can turn on EME without needing to recompile their build; they
just need to create a hidden pref. This will be useful for CDM developers, and
users on platforms where we've not enabled EME yet but users want to test it
(Linux).

We also need to change the GMPUtils.isPluginHidden() accessor so that plugins
are considered hidden if the "visible" pref is false OR we're on an unsupported
platform. This ensures that we must be on a supported OS and the visibility
pref is true before GMPs appear in the addon list.

A consequence of the isPluginHidden() change is that we also need to add a
"force-supported" pref to override the checks that refuse to load the GMPs on
various platform versions, so that the unit tests pass.