bug 1218576 - Ensure IPCTimer is on the main thread. r=froydnj f=gfritzsche
nsTimer fires on the thread that created the timer. An nsTimer instance should
only be manipulated on its target thread (it isn't threadsafe). IPC using
PContent must be on the main thread.

Thus, everything to do with the gIPCTimer must be on the main thread.

This also takes care of bug 1299312.