Bug 1317180, part 2 - Remove really old ~ syntax error for MessageId in IPDL parser. r=billm
MessageId has the production "'~' ID", but if you use it, it produces
an error. This error was added in 2009, in bug 525342. I doubt anybody
expects it to work any more, so it should just be a regular parse
error. This is the only usage of the literal ~ so it can now be
removed from there.