Bug 1176218 - p11. PDM's Supports(TrackInfo) - r=jya
Add a new method alongside SupportsMimeType, which takes a TrackInfo that may
contain more information that just the MIME type.

Eventually the old SupportsMimeType should be removed, but we keep it for now,
as it is used quite extensively and it would be out of scope of this bug to
totally replace it now.