Bug 1223658 - Part 3: Consider fillMode in compositor thread as well. r=birtles
We need to consider fill mode in compositor thread as well as other properties
because pulling the animation back from the compositor thread is sometimes
delayed due to the main thread busyness.  In such situations, if there is
another animation running on the main thread on the same element, users can
easily notice a gap between both of animations.