Bug 1291715 - Add support for 44100 and 48000 Hz sample rates to DtmfInband; r=jesup
The DtmfInband class does not support sample rates above 32000. This adds
support for 44100 and 48000.

The 'a' coefficients were calculated in python as:

int(round(32768*math.cos(2*math.pi*f/fs)))

The 'ym2' coefficients were calculated in python as:

int(round(16383*math.sin(2*math.pi*f/fs)))

where f was in: [697, 770, 852, 941, 1209, 1336, 1477, 1633] and fs was in:
[8000, 16000, 32000, 44100, 44800].

The calculated values were slightly off the existing values at 8000 Hz,
but agreed at 16000 and 32000 Hz.